move helper to api.helpers
